+++ modules/features/note/src/main/java/net/pantasystem/milktea/note/editor/viewmodel/NoteEditorViewModel.kt
@@ -567,159 +567,3 @@ class NoteEditorViewModel @Inject constructor(
 
 
 }
-
-data class NoteEditorFormState(
-    val text: String? = null,
-    val cw: String? = null,
-    val hasCw: Boolean = false,
-)
-
-data class VisibilityAndChannelId(
-    val visibility: Visibility = Visibility.Public(false),
-    val channelId: Channel.Id? = null,
-)
-
-data class NoteEditorSendToState(
-    val visibility: Visibility = Visibility.Public(false),
-    val channelId: Channel.Id? = null,
-    val renoteId: Note.Id? = null,
-    val replyId: Note.Id? = null,
-    val schedulePostAt: Instant? = null,
-    val draftNoteId: Long? = null,
-)
-
-data class NoteEditorUiState(
-    val formState: NoteEditorFormState = NoteEditorFormState(),
-    val sendToState: NoteEditorSendToState = NoteEditorSendToState(),
-    val poll: PollEditingState? = null,
-    val files: List<AppFile> = emptyList(),
-    val currentAccount: Account? = null,
-) {
-    val totalFilesCount: Int
-        get() = this.files.size
-
-    fun checkValidate(textMaxLength: Int = 3000, maxFileCount: Int = 4): Boolean {
-        if (this.files.size > maxFileCount) {
-            return false
-        }
-
-        if ((this.formState.text?.codePointCount(0, this.formState.text.length)
-                ?: 0) > textMaxLength
-        ) {
-            return false
-        }
-
-        if (sendToState.channelId != null && sendToState.visibility != Visibility.Public(true)) {
-            return false
-        }
-
-        if (this.sendToState.renoteId != null) {
-            return true
-        }
-        if (this.poll != null && this.poll.checkValidate()) {
-            return true
-        }
-        return !(
-                this.formState.text.isNullOrBlank()
-                        && this.files.isEmpty()
-                )
-    }
-
-    fun shouldDiscardingConfirmation(): Boolean {
-        val address = (sendToState.visibility as? Visibility.Specified)?.visibleUserIds
-            ?: emptyList()
-        return !formState.text.isNullOrBlank()
-                || files.isNotEmpty()
-                || !poll?.choices.isNullOrEmpty()
-                || address.isNotEmpty()
-    }
-}
-
-
-fun NoteEditorUiState.toCreateNote(account: Account): CreateNote {
-    return CreateNote(
-        author = account,
-        visibility = sendToState.visibility,
-        text = formState.text,
-        cw = if (formState.hasCw) formState.text else null,
-        viaMobile = false,
-        files = files,
-        replyId = sendToState.replyId,
-        renoteId = sendToState.renoteId,
-        poll = poll?.toCreatePoll(),
-        draftNoteId = sendToState.draftNoteId,
-        channelId = sendToState.channelId,
-        scheduleWillPostAt = sendToState.schedulePostAt,
-    )
-}
-
-enum class NoteEditorSavedStateKey() {
-    Text, Cw, PickedFiles, Visibility, ChannelId, ReplyId, RenoteId, ScheduleAt, DraftNoteId, HasCW, Poll
-}
-
-
-fun SavedStateHandle.setText(text: String?) {
-    this[NoteEditorSavedStateKey.Text.name] = text
-}
-
-fun SavedStateHandle.getText(): String? {
-    return this[NoteEditorSavedStateKey.Text.name]
-}
-
-fun SavedStateHandle.setCw(text: String?) {
-    this[NoteEditorSavedStateKey.Cw.name] = text
-}
-
-fun SavedStateHandle.setFiles(files: List<AppFile>) {
-    this[NoteEditorSavedStateKey.PickedFiles.name] = files
-}
-
-fun SavedStateHandle.getFiles(): List<AppFile> {
-    return this[NoteEditorSavedStateKey.PickedFiles.name] ?: emptyList()
-}
-
-fun SavedStateHandle.setChannelId(channelId: Channel.Id?) {
-    this[NoteEditorSavedStateKey.ChannelId.name] = channelId
-}
-
-
-fun SavedStateHandle.setReplyId(noteId: Note.Id?) {
-    this[NoteEditorSavedStateKey.ReplyId.name] = noteId
-}
-
-fun SavedStateHandle.setRenoteId(noteId: Note.Id?) {
-    this[NoteEditorSavedStateKey.RenoteId.name] = noteId
-}
-
-
-fun SavedStateHandle.setScheduleAt(date: Date?) {
-    this[NoteEditorSavedStateKey.ScheduleAt.name] = date
-}
-
-fun SavedStateHandle.setHasCw(hasCw: Boolean) {
-    this[NoteEditorSavedStateKey.HasCW.name] = hasCw
-}
-
-fun SavedStateHandle.getHasCw(): Boolean {
-    return this[NoteEditorSavedStateKey.HasCW.name] ?: false
-}
-
-fun SavedStateHandle.getPoll(): PollEditingState? {
-    return this[NoteEditorSavedStateKey.Poll.name]
-}
-
-fun SavedStateHandle.setPoll(value: PollEditingState?) {
-    this[NoteEditorSavedStateKey.Poll.name] = value
-}
-
-fun SavedStateHandle.setVisibility(visibility: Visibility) {
-    this[NoteEditorSavedStateKey.Visibility.name] = visibility
-}
-
-fun SavedStateHandle.getVisibility(): Visibility {
-    return this[NoteEditorSavedStateKey.Visibility.name] ?: Visibility.Public(false)
-}
-
-fun SavedStateHandle.setDraftNoteId(id: Long?) {
-    this[NoteEditorSavedStateKey.DraftNoteId.name] = id
-}
